Converted class names to pascal case
