+++ app/Models/Farmer/FarmerAddress.php
@@ -1,25 +1,25 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Farmer;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class Farmland extends Model
+class FarmerAddress extends Model
 {
-    public $table = 'farmland';
-
     use HasFactory;
 
     protected $fillable = [
-        'farm_size',
-        'crop_buyer',
+        'house_number',
+        'street',
+        'barangay',
+        'city',
         'created_at',
         'updated_at',
     ];
 
     public function farmer_profile()
     {
-        return $this->hasMany(farmer_profile::class, 'foreign_key');
+        return $this->belongsTo(Farmer_profile::class, 'foreign_key');
     }
 }
