+++ app/Models/Farmer/FarmerProfile.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Farmer;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Orchid\Screen\AsSource;
 use Orchid\Filters\Filterable;
 
-class Farmer_profile extends Model
+class FarmerProfile extends Model
 {
     public $table = 'farmer_profile';
