+++ app/Models/Farmland/Farmland.php
@@ -1,25 +1,25 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Farmland;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class Farmer_address extends Model
+class Farmland extends Model
 {
+    public $table = 'farmland';
+
     use HasFactory;
 
     protected $fillable = [
-        'house_number',
-        'street',
-        'barangay',
-        'city',
+        'farm_size',
+        'crop_buyer',
         'created_at',
         'updated_at',
     ];
 
     public function farmer_profile()
     {
-        return $this->belongsTo(Farmer_profile::class, 'foreign_key');
+        return $this->hasMany(farmer_profile::class, 'foreign_key');
     }
 }
