+++ app/Models/Farmland/FarmlandCropBuyer.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Farmland;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class Farmland_type extends Model
+class FarmlandCropBuyer extends Model
 {
     use HasFactory;
 }
