+++ app/Models/Farmland/FarmlandStatus.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Farmland;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class Farmland_status extends Model
+class FarmlandStatus extends Model
 {
     use HasFactory;
 }
