+++ app/Models/Farmland/FarmlandWateringSystem.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Farmland;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class Farmland_watering_system extends Model
+class FarmlandWateringSystem extends Model
 {
     use HasFactory;
 }
