+++ app/Orchid/Layouts/Farmer/FarmerListLayout.php
@@ -2,7 +2,7 @@
 
 namespace App\Orchid\Layouts\Farmer;
 
-use App\Models\Farmer_profile;
+use App\Models\Farmer\FarmerProfile;
 use Orchid\Screen\Actions\Button;
 use Orchid\Screen\Actions\DropDown;
 use Orchid\Screen\Actions\Link;
@@ -19,7 +19,7 @@ class FarmerListLayout extends Table
      *
      * @var string
      */
-    
+
     protected $target = 'farmer_profile';
 
     /**
@@ -35,43 +35,43 @@ class FarmerListLayout extends Table
                 ->sort()
                 ->cantHide()
                 ->filter(TD::FILTER_TEXT)
-                ->render(function (Farmer_profile $farmer_profile) {
+                ->render(function (FarmerProfile $farmer_profile) {
                     return Link::make($farmer_profile->id)
                         ->route('platform.farmer.profile.edit', $farmer_profile->id);
                 }),
 
             TD::make('lastname', __('Last Name'))
-            ->sort()
-            ->cantHide()
-            ->filter(TD::FILTER_TEXT)
-            ->render(function (Farmer_profile $farmer_profile) {
-                return Link::make($farmer_profile->lastname)
-                    ->route('platform.farmer.profile.edit', $farmer_profile->id);
-            }),
+                ->sort()
+                ->cantHide()
+                ->filter(TD::FILTER_TEXT)
+                ->render(function (FarmerProfile $farmer_profile) {
+                    return Link::make($farmer_profile->lastname)
+                        ->route('platform.farmer.profile.edit', $farmer_profile->id);
+                }),
 
             TD::make('firstname', __('First Name'))
-            ->sort()
-            ->cantHide()
-            ->filter(TD::FILTER_TEXT)
-            ->render(function (Farmer_profile $farmer_profile) {
-                return Link::make($farmer_profile->firstname)
-                    ->route('platform.farmer.profile.edit', $farmer_profile->id);
-            }),
+                ->sort()
+                ->cantHide()
+                ->filter(TD::FILTER_TEXT)
+                ->render(function (FarmerProfile $farmer_profile) {
+                    return Link::make($farmer_profile->firstname)
+                        ->route('platform.farmer.profile.edit', $farmer_profile->id);
+                }),
 
             TD::make('middlename', __('Middle Name'))
-            ->sort()
-            ->cantHide()
-            ->filter(TD::FILTER_TEXT)
-            ->render(function (Farmer_profile $farmer_profile) {
-                return Link::make($farmer_profile->middlename)
-                    ->route('platform.farmer.profile.edit', $farmer_profile->id);
-            }),
+                ->sort()
+                ->cantHide()
+                ->filter(TD::FILTER_TEXT)
+                ->render(function (FarmerProfile $farmer_profile) {
+                    return Link::make($farmer_profile->middlename)
+                        ->route('platform.farmer.profile.edit', $farmer_profile->id);
+                }),
 
             TD::make(__('Actions'))
                 ->align(TD::ALIGN_CENTER)
                 ->cantHide()
                 ->width('100px')
-                ->render(function (Farmer_profile $farmer_profile) {
+                ->render(function (FarmerProfile $farmer_profile) {
                     return DropDown::make()
                         ->icon('options-vertical')
                         ->list([
@@ -87,7 +87,7 @@ class FarmerListLayout extends Table
                                     'id' => $farmer_profile->id,
                                 ]),
                         ]);
-            }),
+                }),
         ];
     }
 }
