+++ app/Orchid/Screens/Farmer/FarmerEditScreen.php
@@ -3,7 +3,7 @@
 namespace App\Orchid\Screens\Farmer;
 
 use Illuminate\Http\Request;
-use App\Models\Farmer_profile;
+use App\Models\Farmer\FarmerProfile;
 use App\Orchid\Layouts\Farmer\FarmerCreateLoginLayout;
 use App\Orchid\Layouts\Farmer\FarmerCreateProfileLayout;
 use App\Orchid\Layouts\Farmer\FarmerCreateSkillLayout;
@@ -43,10 +43,10 @@ class FarmerEditScreen extends Screen
      * @return array
      */
 
-    public function query(Farmer_profile $farmer_profile): array
+    public function query(FarmerProfile $farmer_profile): array
     {
         $this->farmer_profile = $farmer_profile;
-        
+
         if (!$farmer_profile->exists) {
             $this->name = 'Enroll Farmer';
             $this->description = 'Enroll New Farmer';
@@ -90,7 +90,7 @@ class FarmerEditScreen extends Screen
             Layout::block(FarmerCreateProfileLayout::class)
                 ->title('Personal Information')
                 ->description("This information collects farmer's personal information."),
-            
+
             Layout::block(FarmerCreateAddressLayout::class)
                 ->title('Personal Address')
                 ->description("This information collects farmer's personal address."),
@@ -106,13 +106,13 @@ class FarmerEditScreen extends Screen
     }
 
     /**
-     * @param Farmer_profile    $farmer_profile
+     * @param FarmerProfile    $farmer_profile
      * @param Request $request
      *
      * @return \Illuminate\Http\RedirectResponse
      */
 
-    public function save(Farmer_profile $farmer_profile, Request $request)
+    public function save(FarmerProfile $farmer_profile, Request $request)
     {
         $request->validate([
             'farmer_profile.gender' => [
@@ -205,14 +205,14 @@ class FarmerEditScreen extends Screen
     }
 
     /**
-     * @param Farmer_profile $farmer_profile
+     * @param FarmerProfile $farmer_profile
      *
      * @throws \Exception
      *
      * @return \Illuminate\Http\RedirectResponse
      */
 
-    public function remove(Farmer_profile $farmer_profile)
+    public function remove(FarmerProfile $farmer_profile)
     {
         $farmer_profile->delete();
